js/src/xpconnect/tests/mochitest/test_bug650273.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=650273
test by moz_bug_r_a4@yahoo.com
-->
<head>
  <title>Test for Bug 650273</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank"
href="https://bugzilla.mozilla.org/show_bug.cgi?id=650273">Mozilla Bug 650273</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 504877 **/
SimpleTest.waitForExplicitFinish();
var count = 0;
function done() {
    if (++count == 2) {
        try {
            ok($('ifr').location.host === 'example.com', "shouldn't see this");
        } catch (e) {
            ok(true, "navigation successfully happened");
        }
        SimpleTest.finish();
    }
}

</script>

<iframe id="ifr" src="file_bug650273.html" onload="done()"></iframe>

</pre>
</body>
</html>
